groups typings and flattens file structure
