--- libs/menu-matriarch/types/src/lib/date-name-display.type.ts
@@ -1,1 +1,0 @@
-export type DayNameDisplay = 'full';
