--- libs/menu-matriarch/types/src/lib/dish-type.type.ts
@@ -1,1 +1,0 @@
-export type DishType = 'main' | 'side' | 'dessert';
