--- libs/menu-matriarch/types/src/lib/dtos/meal-dto.interface.ts
@@ -1,8 +1,0 @@
-export interface MealDto {
-  id: string;
-  uid: string;
-  name: string;
-  description: string;
-  dishIds: string[];
-  tagIds: string[];
-}
