--- libs/menu-matriarch/types/src/lib/dtos/menu-dto.interface.ts
@@ -1,12 +1,0 @@
-import { Day } from '../day.type';
-
-export interface MenuDto {
-  id: string;
-  uid: string;
-  name: string;
-  favorited: boolean;
-  startDay: Day;
-  contents: {
-    [day in Day]: string[];
-  };
-}
