--- libs/menu-matriarch/types/src/lib/dtos/user-dto.interface.ts
@@ -1,8 +1,0 @@
-import { UserPreferences } from '../user-preferences.interface';
-
-export interface UserDto {
-  uid: string;
-  name: string;
-  email: string;
-  preferences: UserPreferences;
-}
