--- libs/menu-matriarch/types/src/lib/filtered-dishes.interface.ts
@@ -1,8 +1,0 @@
-import { DishType } from './dish-type.type';
-import { Dish } from './entities/dish.interface';
-
-export interface FilteredDishesGroup {
-  type: DishType;
-  dishes: Dish[];
-  placeholderText?: string;
-}
