+++ libs/menu-matriarch/types/src/lib/meal.interface.ts
@@ -1,6 +1,14 @@
-import { MealDto } from '../dtos/meal-dto.interface';
 import { Dish } from './dish.interface';
 import { Tag } from './tag.interface';
+
+export interface MealDto {
+  id: string;
+  uid: string;
+  name: string;
+  description: string;
+  dishIds: string[];
+  tagIds: string[];
+}
 
 export interface Meal extends Omit<MealDto, 'dishIds' | 'tagIds'> {
   dishes: Dish[];
