--- libs/menu-matriarch/types/src/lib/menu-entry.interface.ts
@@ -1,7 +1,0 @@
-import { Day } from './day.type';
-import { Dish } from './entities/dish.interface';
-
-export interface MenuEntry {
-  day: Day;
-  dishes: Dish[];
-}
