--- libs/menu-matriarch/types/src/lib/tag-model.interface.ts
@@ -1,5 +1,0 @@
-import { Tag } from './entities/tag.interface';
-
-export interface TagModel extends Tag {
-  checked: boolean;
-}
