+++ libs/menu-matriarch/types/src/lib/tag.interface.ts
@@ -6,3 +6,9 @@ export interface TagDto {
   mealIds: string[];
   dishIds: string[];
 }
+
+export type Tag = TagDto;
+
+export interface TagModel extends Tag {
+  checked: boolean;
+}
